Test Walker in a separate module
