+++ concert/tests/integration/test_experiments.py
@@ -10,8 +10,8 @@ from concert.experiments.imaging import (Experiment as ImagingExperiment,
                                          tomo_angular_step, tomo_max_speed,
                                          tomo_projections_number, frames)
 from concert.devices.cameras.dummy import Camera
-from concert.storage import Walker
 from concert.tests import TestCase, suppressed_logging, assert_almost_equal
+from concert.tests.unit.test_walker import DummyWalker, compute_path
 
 
 @suppressed_logging
@@ -45,42 +45,6 @@ def test_frames():
     count.i = 0
     inject(frames(5, Camera()), count())
     assert count.i == 5
-
-
-def compute_path(*parts):
-    return '/'.join(parts)
-
-
-class DummyWalker(Walker):
-    def __init__(self, root=''):
-        super(DummyWalker, self).__init__(root)
-        self._paths = set([])
-
-    @property
-    def paths(self):
-        return self._paths
-
-    def exists(self, *paths):
-        return compute_path(*paths) in self._paths
-
-    def _descend(self, name):
-        self._current = compute_path(self._current, name)
-        self._paths.add(self._current)
-
-    def _ascend(self):
-        if self._current != self._root:
-            self._current = compute_path(*self._current.split('/')[:-1])
-
-    @coroutine
-    def _write_coroutine(self, fname=None):
-        fname = fname if fname is not None else self._fname
-        path = compute_path(self._current, fname)
-
-        i = 0
-        while True:
-            yield
-            self._paths.add(compute_path(path, str(i)))
-            i += 1
 
 
 class TestExperimentBase(TestCase):
